+++ src/main/java/lotto/domain/dto/WinningResult.java
@@ -2,7 +2,7 @@ package lotto.domain.dto;
 
 import java.util.Arrays;
 import java.util.LinkedHashMap;
-import lotto.constant.Prize;
+import lotto.domain.Prize;
 
 public record WinningResult(LinkedHashMap<Prize, Integer> prizes) {
     public static WinningResult of(Prize... prizes) {
